Changed vector field option to arc length option
I reviewed how to calculate the integral for a vector field, and it seemed
more complicated than what I want to deal with for this application.
Computing an arc length integral works, because I can avoid deriving by
starting with the derivative of the position vector function, even though
that's unusual.